+++ src/bin/run-3x.ts
@@ -2,7 +2,7 @@ import 'source-map-support/register';
 import { engineConfigFromENV } from '../lib/config/env';
 import DataDir from '../lib/data/dir';
 import { Data, DataSet } from '../lib/data/set';
-import { Engine } from "../lib/engine/engine";
+import { Engine } from "../lib/engine";
 import { Hubspot } from '../lib/hubspot';
 import { Logger } from '../lib/log';
