+++ src/lib/engine/summary.ts
@@ -1,8 +1,8 @@
+import { Engine } from ".";
 import { Table } from "../log/table";
 import { Deal } from "../model/deal";
 import { formatMoney, formatNumber } from "../util/formatters";
 import { isPresent } from "../util/helpers";
-import { Engine } from "./engine";
 
 export function printSummary(engine: Engine) {
